+++ packages/atomic/src/utils/custom-element-tags.ts
@@ -7,7 +7,7 @@
  * Stencil components are excluded as they lazy-load.
  */
 export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
-  'answer-content',
+  'atomic-answer-content',
   'atomic-aria-live',
   'atomic-automatic-facet',
   'atomic-automatic-facet-generator',
@@ -58,6 +58,8 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-format-unit',
   'atomic-generated-answer',
   'atomic-generated-answer-feedback-modal',
+  'atomic-generated-answer-thread-item',
+  'atomic-generated-answers-thread',
   'atomic-html',
   'atomic-icon',
   'atomic-insight-edit-toggle',
@@ -65,6 +67,7 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-insight-folded-result-list',
   'atomic-insight-full-search-button',
   'atomic-insight-generate-answer-button',
+  'atomic-insight-generated-answer',
   'atomic-insight-history-toggle',
   'atomic-insight-interface',
   'atomic-insight-layout',
@@ -76,9 +79,12 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-insight-refine-modal',
   'atomic-insight-refine-toggle',
   'atomic-insight-result',
+  'atomic-insight-result-action',
   'atomic-insight-result-action-bar',
+  'atomic-insight-result-attach-to-case-action',
   'atomic-insight-result-attach-to-case-indicator',
   'atomic-insight-result-children',
+  'atomic-insight-result-children-template',
   'atomic-insight-result-list',
   'atomic-insight-result-quickview-action',
   'atomic-insight-result-template',
@@ -88,6 +94,7 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-insight-smart-snippet-suggestions',
   'atomic-insight-tab',
   'atomic-insight-tabs',
+  'atomic-insight-timeframe-facet',
   'atomic-insight-user-actions-modal',
   'atomic-insight-user-actions-session',
   'atomic-insight-user-actions-timeline',
@@ -95,6 +102,8 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-ipx-button',
   'atomic-ipx-embedded',
   'atomic-ipx-modal',
+  'atomic-ipx-recs-list',
+  'atomic-ipx-refine-modal',
   'atomic-ipx-refine-toggle',
   'atomic-ipx-result-link',
   'atomic-ipx-tab',
@@ -201,7 +210,6 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-text',
   'atomic-timeframe',
   'atomic-timeframe-facet',
-  'generated-answer-thread-item',
 ]);
 
 /**